od5.0: Fixed conversion warning
Summary:
Fixed:
src/osgGeo/TiledOffScreenRenderer.cpp: In member function ‘void osgGeo::TiledOffScreenRenderer::setOutputSize(int, int)’:
src/osgGeo/TiledOffScreenRenderer.cpp:72: warning: converting to ‘int’ from ‘double’
src/osgGeo/TiledOffScreenRenderer.cpp:73: warning: converting to ‘int’ from ‘double’

also applied a few other cosmetical changes from od to make the file identical.

Test Plan: Not tested.

Reviewers: jaap, ding

Reviewed By: ding

Differential Revision: https://phabricator.dgbes.com:8443/D5307

+++ src/osgGeo/LayeredTexture.cpp
@@ -1573,7 +1573,7 @@ float LayeredTexture::getMaxAnisotropy( int layerIdx ) const
     
     float maxAnisotropy = powerOf2Ceil( (int) floor(ratio+0.5) );
 
-    int power = _anisotropicPower;
+    int power = (int)_anisotropicPower;
     while ( (--power)>=0 )
 	maxAnisotropy *= 2.0;
 
@@ -1974,7 +1974,6 @@ void LayeredTexture::updateSetupStateSetIfNeeded()
     {
 	if ( !_retileCompositeLayer )
 	    _compositeLayerUpdate = true;
-
 	buildShaders();
 	setUpdateVar( _updateSetupStateSet, false );
     }
